requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
motor==3.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
python-telegram-bot==21.3
apscheduler==3.10.4
slowapi==0.1.9
pymongo==4.7.2
dnspython==2.6.1
httpx==0.27.0
